refactor: move querydb.language.android to semanticcpg
background: we'd like to use it downstream without inheriting a large
dependency tree incl. console and various frontends
