+++ semanticcpg/src/main/scala/io/shiftleft/semanticcpg/language/android/LocalTraversal.scala
@@ -1,4 +1,4 @@
-package io.joern.querydb.language.android
+package io.shiftleft.semanticcpg.language.android
 
 import io.shiftleft.codepropertygraph.generated.nodes.Local
 import io.shiftleft.semanticcpg.language._
